rss-scraper/requirements.txt
# RSS Scraper Dependencies
feedparser>=6.0.10
httpx>=0.25.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.53.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Testing Dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.11.0
pytest-cov>=4.1.0
httpx-mock>=0.11.0

# Development Dependencies
black>=23.0.0
ruff>=0.1.0
mypy>=1.5.0